Moved wal out of raftex, so that it can be used by other modules
