+++ src/CMakeLists.txt
@@ -3,6 +3,7 @@ add_subdirectory(interface)
 add_subdirectory(client)
 add_subdirectory(meta)
 add_subdirectory(console)
+add_subdirectory(wal)
 add_subdirectory(raftex)
 add_subdirectory(kvstore)
 add_subdirectory(parser)
