+++ src/raftex/test/RaftexTestBase.h
@@ -17,16 +17,21 @@
 
 
 namespace nebula {
+
+namespace wal {
+class BufferFlusher;
+}  // namespace wal
+
+
 namespace raftex {
 
 class RaftexService;
-class BufferFlusher;
 
 namespace test {
 class TestShard;
 }  // namespace test
 
-extern std::unique_ptr<BufferFlusher> flusher;
+extern std::unique_ptr<wal::BufferFlusher> flusher;
 
 extern std::mutex leaderMutex;
 extern std::condition_variable leaderCV;
