+++ src/raftex/test/TestShard.h
@@ -25,7 +25,7 @@ public:
         PartitionID partId,
         HostAddr addr,
         const folly::StringPiece walRoot,
-        BufferFlusher* flusher,
+        wal::BufferFlusher* flusher,
         std::shared_ptr<folly::IOThreadPoolExecutor> ioPool,
         std::shared_ptr<thread::GenericThreadPool> workers,
         std::function<void (size_t idx, const char*, TermID)>
